feat: current ingredients refactored
